Add ability to specify samples directory
Modified the check.sh script to allow the user to specify the samples directory
as the first positional parameter to the script. By default, it will choose the
directory named 'samples'.

For example to run against all tests in samples-checkpoint directory:

        ./check samples-checkpoint

To run against only t1, t2, and t3 in samples-checkpoint directory:

        ./check samples-checkpoint t1 t2 t3

To run agains only t1, t2, and t3 in samples directory:

        ./check samples t1 t2 t3