Begin rework of scope checking implementation
The Scope checking implementation has gotten fairly ugly and hard to follow.
Thus, it needs to be reworked and reimplemented in a way that makes it easier
to understand and extend. This commit marks the start of that effort.

The first pass of the compiler will call the Program's BuildScope. This pass
will build up the initial scope information, creating the global symbol table
and the sub scope's symbol tables.