Reordered the placement of errors in bad6.out
Since my scanner outputs the same errors found in bad6.out, but in a different
order, the .out file was modified. The semantic checker is still valid if it
reports the errors in a different order. The important thing is that it reports
the errors that it expects it to report (order is not important).