Add GetType method for FieldAccess class
Added support for getting the type of FieldAccess instances that have NULL
bases. In order to implement this, the CompoundExpr had to be extended to
provide a custom BuildScope method. Also, fixed a bug in a couple of the Expr's
Check methods; they need to each call their member's respective Check method.