Move bulk of FieldAccess methods into Expr class
Moved the bulk effort of the FieldAccess methods into the Expr class. This
allows other subclasses to make use of this functionality. Also attempted to
fix the functions to properly climb an object's inheritance hierarchy.
Previously, a field's declaration would not be found if it was declared in a
base class and accessed in a derived class.