Scope instances now hold pointer to ClassDecl
Modified the Scope class to hold a ClassDecl pointer if it is a sub scope for a
class. This allows the current class scope's declaration to be obtained simply
by traversing up the scope tree hierarchy. The related Expr::IsInClassScope
method was changed to Expr::GetClassScope to reflect this change.