Fixed a bug found in GetType and Check methods
The FieldAccess class's GetType and Check methods along with the Call class's
Check method were not working correctly if base was NULL but the field was an
inherited member from another class. Previously, these methods would not check
the inherited class's scope for a declaration for the field in question.